pyproject.toml
[project]
name = "data-dedupliacation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "nltk>=3.9.2",
    "numpy<2.0.0",
    "pytest>=9.0.2",
    "rich>=14.3.2",
    "ruff>=0.15.0",
    "typer>=0.9.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "lxml>=6.0.2",
    "xmltodict>=1.0.4",
    "scikit-learn>=1.8.0",
    "xxhash>=3.6.0",
    "fastembed>=0.3.4",
    "javalang>=0.13.0",
    "datasketch>=1.9.0",
    "pandas>=2.3.3",
]
[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'"
]
